backend/requirements.txt
# Generated from current venv via pip freeze
# Security: pin Starlette to patched version to address CVE (see Dependabot/pip-audit)
aiofiles==25.1.0
alembic==1.17.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
bcrypt==5.0.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.120.1
filelock==3.20.0
fsspec==2025.9.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
Mako==1.3.10
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.4
openai==2.6.1
packaging==25.0
passlib==1.7.4
pillow==12.0.0
pluggy==1.6.0
psutil==7.1.3
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
PyMuPDF==1.26.5
pytest==8.4.2
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
# Knowledge Base Dependencies
faiss-cpu==1.12.0
PyPDF2==3.0.1
python-docx==1.1.2
lxml==6.0.2
# Futuristic Knowledge Base Features
rank-bm25==0.2.2  # BM25 keyword search for hybrid search
regex==2025.10.23
requests==2.32.5
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
